Clean-up and comments

+++ ev.h
@@ -1,8 +1,10 @@
+//macros to enable and disable event flags
+
 #define EV_ENABLE(ev_enum) (ev_flags |= (1 << ev_enum))
 #define EV_DISABLE(ev_enum) (ev_flags &= ~(1 << ev_enum))
 #define IF_EV(ev_num) if(ev_flags & (1 << ev_num))
 
-extern unsigned long ev_flags;
+extern unsigned int ev_flags;
 
 typedef enum Event {
 	ev_ciao,
